send email to confirm user has been added to group following T&C acceptation
